+++ src/Settings/StorageHandler.php
@@ -129,6 +129,9 @@ class StorageHandler
         '67316', // X-IG-BLOKS-SERIALIZE-PAYLOAD
         '76504', // X-Bloks-Prism-Button-Version 0x81117D00003FD1
         '62246', // X-Bloks-Prism-Colors-Enabled 0x81093600052121
+        '55958', // banyan. direct_ibc_inbox_discovery and direct_ibc_inbox_invitations
+        '69705', // banyan. forwarding_recipient_sheet and story_share_sheet
+        '68215', // banyan. Banyan view selector
     ];
 
     /**